add styles to the signup component
