+++ src/pages/index.js
@@ -6,7 +6,6 @@ import Hero from "../components/Hero";
 import Help from "../components/Help";
 import Testimonials from "../components/Testimonials";
 import WhyUs from "../components/WhyUs";
-import SignUp from "../components/SignUp";
 
 export default function Home() {
   const { t } = useTranslation("index");
@@ -17,13 +16,12 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      {/* <Hero />
+      <Hero />
       <CourseCards />
       <InterestCard />
       <WhyUs />
       <Testimonials />
-      <Help /> */}
-      <SignUp />
+      <Help />
     </div>
   );
 }
